ui: cache file stats at (de)index phase
Instead of calculating the stats on page load and caching it for an
amount of time, this change calculates the stats whenever a build gets
indexed or deindexed (in the worker task) then caches the result. This
ensures that a) the cached data is always current and b) calculating the
stats is not required at page load.

This also intruduces data structures for the collected stats.