gateway: add register_signal
Add a register_signal method to the gateway. The gbp_fl plugin itself
will want to register it's own signals with the GBP signal dispatcher.
For example, the tasks which are presently updaing the UI stats cache
themselves will be able to de-couple from the UI by emitting signals and
letting a signal handler take care of the stats caching.