add plugin checks
Since the builds and file records have the possibility of getting out of
sync, this utilizes the new "plugin checks" feature of Gentoo Build
Publisher to add to checks: one to ensure that every GBP Build has file
records and one to ensure that every file record has a corresponding GBP
Build. These checks return warnings instead of errors because when there
is an inconsistency it's not fatal.