graphql: remove BinPkg type from the schema
Instead of having a "custom" type for the package, simply resolve
Package.

There are still the internal gbp_fl.types that are basically mirrors of
the types in GBP. Eventually these will be removed as well and replaced
with their GBP counterparts. The "gateway" interface seemed like a good
idea at the time, but in practice it's just a lot of extra coding around
tye GBP types with are public interfaces. It's easier/better to use the
public types and extend them in the GraphQL interface as needed.